Refactor examples to work with tutorial backend
