+++ databuilder/ehrql-tutorial-examples/2a_multiple_dataset_definition.py
@@ -1,5 +1,6 @@
 from databuilder.query_language import Dataset
-from databuilder.tables import patients, prescriptions
+from databuilder.tables.beta.tutorial import patient_demographics as patients
+from databuilder.tables.beta.tutorial import prescriptions
 
 
 dataset = Dataset()
